requirements.txt
openai
langchain
langgraph
pydantic
fastapi
uvicorn
python-dotenv
pandas
numpy
pymupdf
pdf2image
pillow
pytesseract
easyocr
twilio
gtts
python-multipart
google-genai
fpdf2